requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
colorama==0.4.6
defusedxml==0.7.1
fonttools==4.61.1
fpdf2==2.8.5
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==1.2.0
logistro==2.0.1
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.4.2
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.2
pluggy==1.6.0
protobuf==6.33.5
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
streamlit==1.54.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
xlsxwriter==3.2.9
